requirements.txt
pip install the following packages:
PyGithub
python-dotenv
pillow
platform
gitpython